// src/generated/services/Jia_ll_demployeesService.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */

import type { GetEntityMetadataOptions, EntityMetadata } from '@microsoft/power-apps/data/metadata/dataverse';
import type { IGetOptions, IGetAllOptions } from '../models/CommonModels';
import type { IOperationResult } from '@microsoft/power-apps/data';
import type { Jia_ll_demployeesBase, Jia_ll_demployees } from '../models/Jia_ll_demployeesModel';
import { dataSourcesInfo } from '../../../.power/schemas/appschemas/dataSourcesInfo';
import { getClient } from '@microsoft/power-apps/data';


export class Jia_ll_demployeesService {
  private static readonly dataSourceName = 'jia_ll_demployees';

  private static readonly client = getClient(dataSourcesInfo);

  public static async create(record: Omit<Jia_ll_demployeesBase, 'jia_ll_demployeeid'>): Promise<IOperationResult<Jia_ll_demployees>> {
    const result = await Jia_ll_demployeesService.client.createRecordAsync<Omit<Jia_ll_demployeesBase, 'jia_ll_demployeeid'>, Jia_ll_demployees>(
      Jia_ll_demployeesService.dataSourceName,
      record
    );
    return result;
  }

  public static async update(id: string, changedFields: Partial<Omit<Jia_ll_demployeesBase, 'jia_ll_demployeeid'>>): Promise<IOperationResult<Jia_ll_demployees>> {
    const result = await Jia_ll_demployeesService.client.updateRecordAsync<Partial<Omit<Jia_ll_demployeesBase, 'jia_ll_demployeeid'>>, Jia_ll_demployees>(
      Jia_ll_demployeesService.dataSourceName,
      id.toString(),
      changedFields
    );
    return result;
  }

  public static async delete(id: string): Promise<void> {
    await Jia_ll_demployeesService.client.deleteRecordAsync(
      Jia_ll_demployeesService.dataSourceName,
      id.toString());
  }

  public static async get(id: string, options?: IGetOptions): Promise<IOperationResult<Jia_ll_demployees>> {
    const result = await Jia_ll_demployeesService.client.retrieveRecordAsync<Jia_ll_demployees>(
      Jia_ll_demployeesService.dataSourceName,
      id.toString(),
      options
    );
    return result;
  }

  public static async getAll(options?: IGetAllOptions): Promise<IOperationResult<Jia_ll_demployees[]>> {
    const result = await Jia_ll_demployeesService.client.retrieveMultipleRecordsAsync<Jia_ll_demployees>(
      Jia_ll_demployeesService.dataSourceName,
      options
    );
    return result;
  }

  public static getMetadata(
    options: GetEntityMetadataOptions<Jia_ll_demployees> = {}
  ): Promise<IOperationResult<Partial<EntityMetadata>>> {
    return Jia_ll_demployeesService.client.executeAsync({
      dataverseRequest: {
        action: "getEntityMetadata",
        parameters: {
          tableName: Jia_ll_demployeesService.dataSourceName,
          options: options as GetEntityMetadataOptions,
        },
      },
    });
  }
}