// src/generated/services/Jia_ll_dshiftgroupsService.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */

import type { GetEntityMetadataOptions, EntityMetadata } from '@microsoft/power-apps/data/metadata/dataverse';
import type { IGetOptions, IGetAllOptions } from '../models/CommonModels';
import type { IOperationResult } from '@microsoft/power-apps/data';
import type { Jia_ll_dshiftgroupsBase, Jia_ll_dshiftgroups } from '../models/Jia_ll_dshiftgroupsModel';
import { dataSourcesInfo } from '../../../.power/schemas/appschemas/dataSourcesInfo';
import { getClient } from '@microsoft/power-apps/data';


export class Jia_ll_dshiftgroupsService {
  private static readonly dataSourceName = 'jia_ll_dshiftgroups';

  private static readonly client = getClient(dataSourcesInfo);

  public static async create(record: Omit<Jia_ll_dshiftgroupsBase, 'jia_ll_dshiftgroupid'>): Promise<IOperationResult<Jia_ll_dshiftgroups>> {
    const result = await Jia_ll_dshiftgroupsService.client.createRecordAsync<Omit<Jia_ll_dshiftgroupsBase, 'jia_ll_dshiftgroupid'>, Jia_ll_dshiftgroups>(
      Jia_ll_dshiftgroupsService.dataSourceName,
      record
    );
    return result;
  }

  public static async update(id: string, changedFields: Partial<Omit<Jia_ll_dshiftgroupsBase, 'jia_ll_dshiftgroupid'>>): Promise<IOperationResult<Jia_ll_dshiftgroups>> {
    const result = await Jia_ll_dshiftgroupsService.client.updateRecordAsync<Partial<Omit<Jia_ll_dshiftgroupsBase, 'jia_ll_dshiftgroupid'>>, Jia_ll_dshiftgroups>(
      Jia_ll_dshiftgroupsService.dataSourceName,
      id.toString(),
      changedFields
    );
    return result;
  }

  public static async delete(id: string): Promise<void> {
    await Jia_ll_dshiftgroupsService.client.deleteRecordAsync(
      Jia_ll_dshiftgroupsService.dataSourceName,
      id.toString());
  }

  public static async get(id: string, options?: IGetOptions): Promise<IOperationResult<Jia_ll_dshiftgroups>> {
    const result = await Jia_ll_dshiftgroupsService.client.retrieveRecordAsync<Jia_ll_dshiftgroups>(
      Jia_ll_dshiftgroupsService.dataSourceName,
      id.toString(),
      options
    );
    return result;
  }

  public static async getAll(options?: IGetAllOptions): Promise<IOperationResult<Jia_ll_dshiftgroups[]>> {
    const result = await Jia_ll_dshiftgroupsService.client.retrieveMultipleRecordsAsync<Jia_ll_dshiftgroups>(
      Jia_ll_dshiftgroupsService.dataSourceName,
      options
    );
    return result;
  }

  public static getMetadata(
    options: GetEntityMetadataOptions<Jia_ll_dshiftgroups> = {}
  ): Promise<IOperationResult<Partial<EntityMetadata>>> {
    return Jia_ll_dshiftgroupsService.client.executeAsync({
      dataverseRequest: {
        action: "getEntityMetadata",
        parameters: {
          tableName: Jia_ll_dshiftgroupsService.dataSourceName,
          options: options as GetEntityMetadataOptions,
        },
      },
    });
  }
}