// src/generated/services/Jia_ll_dshiftplansService.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */

import type { GetEntityMetadataOptions, EntityMetadata } from '@microsoft/power-apps/data/metadata/dataverse';
import type { IGetOptions, IGetAllOptions } from '../models/CommonModels';
import type { IOperationResult } from '@microsoft/power-apps/data';
import type { Jia_ll_dshiftplansBase, Jia_ll_dshiftplans } from '../models/Jia_ll_dshiftplansModel';
import { dataSourcesInfo } from '../../../.power/schemas/appschemas/dataSourcesInfo';
import { getClient } from '@microsoft/power-apps/data';


export class Jia_ll_dshiftplansService {
  private static readonly dataSourceName = 'jia_ll_dshiftplans';

  private static readonly client = getClient(dataSourcesInfo);

  public static async create(record: Omit<Jia_ll_dshiftplansBase, 'jia_ll_dshiftplanid'>): Promise<IOperationResult<Jia_ll_dshiftplans>> {
    const result = await Jia_ll_dshiftplansService.client.createRecordAsync<Omit<Jia_ll_dshiftplansBase, 'jia_ll_dshiftplanid'>, Jia_ll_dshiftplans>(
      Jia_ll_dshiftplansService.dataSourceName,
      record
    );
    return result;
  }

  public static async update(id: string, changedFields: Partial<Omit<Jia_ll_dshiftplansBase, 'jia_ll_dshiftplanid'>>): Promise<IOperationResult<Jia_ll_dshiftplans>> {
    const result = await Jia_ll_dshiftplansService.client.updateRecordAsync<Partial<Omit<Jia_ll_dshiftplansBase, 'jia_ll_dshiftplanid'>>, Jia_ll_dshiftplans>(
      Jia_ll_dshiftplansService.dataSourceName,
      id.toString(),
      changedFields
    );
    return result;
  }

  public static async delete(id: string): Promise<void> {
    await Jia_ll_dshiftplansService.client.deleteRecordAsync(
      Jia_ll_dshiftplansService.dataSourceName,
      id.toString());
  }

  public static async get(id: string, options?: IGetOptions): Promise<IOperationResult<Jia_ll_dshiftplans>> {
    const result = await Jia_ll_dshiftplansService.client.retrieveRecordAsync<Jia_ll_dshiftplans>(
      Jia_ll_dshiftplansService.dataSourceName,
      id.toString(),
      options
    );
    return result;
  }

  public static async getAll(options?: IGetAllOptions): Promise<IOperationResult<Jia_ll_dshiftplans[]>> {
    const result = await Jia_ll_dshiftplansService.client.retrieveMultipleRecordsAsync<Jia_ll_dshiftplans>(
      Jia_ll_dshiftplansService.dataSourceName,
      options
    );
    return result;
  }

  public static getMetadata(
    options: GetEntityMetadataOptions<Jia_ll_dshiftplans> = {}
  ): Promise<IOperationResult<Partial<EntityMetadata>>> {
    return Jia_ll_dshiftplansService.client.executeAsync({
      dataverseRequest: {
        action: "getEntityMetadata",
        parameters: {
          tableName: Jia_ll_dshiftplansService.dataSourceName,
          options: options as GetEntityMetadataOptions,
        },
      },
    });
  }
}