// src/generated/services/Jia_ll_fattendancereocrdsService.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */

import type { GetEntityMetadataOptions, EntityMetadata } from '@microsoft/power-apps/data/metadata/dataverse';
import type { IGetOptions, IGetAllOptions } from '../models/CommonModels';
import type { IOperationResult } from '@microsoft/power-apps/data';
import type { Jia_ll_fattendancereocrdsBase, Jia_ll_fattendancereocrds } from '../models/Jia_ll_fattendancereocrdsModel';
import { dataSourcesInfo } from '../../../.power/schemas/appschemas/dataSourcesInfo';
import { getClient } from '@microsoft/power-apps/data';


export class Jia_ll_fattendancereocrdsService {
  private static readonly dataSourceName = 'jia_ll_fattendancereocrds';

  private static readonly client = getClient(dataSourcesInfo);

  public static async create(record: Omit<Jia_ll_fattendancereocrdsBase, 'jia_ll_fattendancereocrdid'>): Promise<IOperationResult<Jia_ll_fattendancereocrds>> {
    const result = await Jia_ll_fattendancereocrdsService.client.createRecordAsync<Omit<Jia_ll_fattendancereocrdsBase, 'jia_ll_fattendancereocrdid'>, Jia_ll_fattendancereocrds>(
      Jia_ll_fattendancereocrdsService.dataSourceName,
      record
    );
    return result;
  }

  public static async update(id: string, changedFields: Partial<Omit<Jia_ll_fattendancereocrdsBase, 'jia_ll_fattendancereocrdid'>>): Promise<IOperationResult<Jia_ll_fattendancereocrds>> {
    const result = await Jia_ll_fattendancereocrdsService.client.updateRecordAsync<Partial<Omit<Jia_ll_fattendancereocrdsBase, 'jia_ll_fattendancereocrdid'>>, Jia_ll_fattendancereocrds>(
      Jia_ll_fattendancereocrdsService.dataSourceName,
      id.toString(),
      changedFields
    );
    return result;
  }

  public static async delete(id: string): Promise<void> {
    await Jia_ll_fattendancereocrdsService.client.deleteRecordAsync(
      Jia_ll_fattendancereocrdsService.dataSourceName,
      id.toString());
  }

  public static async get(id: string, options?: IGetOptions): Promise<IOperationResult<Jia_ll_fattendancereocrds>> {
    const result = await Jia_ll_fattendancereocrdsService.client.retrieveRecordAsync<Jia_ll_fattendancereocrds>(
      Jia_ll_fattendancereocrdsService.dataSourceName,
      id.toString(),
      options
    );
    return result;
  }

  public static async getAll(options?: IGetAllOptions): Promise<IOperationResult<Jia_ll_fattendancereocrds[]>> {
    const result = await Jia_ll_fattendancereocrdsService.client.retrieveMultipleRecordsAsync<Jia_ll_fattendancereocrds>(
      Jia_ll_fattendancereocrdsService.dataSourceName,
      options
    );
    return result;
  }

  public static getMetadata(
    options: GetEntityMetadataOptions<Jia_ll_fattendancereocrds> = {}
  ): Promise<IOperationResult<Partial<EntityMetadata>>> {
    return Jia_ll_fattendancereocrdsService.client.executeAsync({
      dataverseRequest: {
        action: "getEntityMetadata",
        parameters: {
          tableName: Jia_ll_fattendancereocrdsService.dataSourceName,
          options: options as GetEntityMetadataOptions,
        },
      },
    });
  }
}