// src/generated/models/Jia_ll_demployeesModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */
export const Jia_ll_demployeesstatecode = {
  0: 'Active',
  1: 'Inactive'
} as const;
export type Jia_ll_demployeesstatecode = keyof typeof Jia_ll_demployeesstatecode;
export const Jia_ll_demployeesstatuscode = {
  1: 'Active',
  2: 'Inactive'
} as const;
export type Jia_ll_demployeesstatuscode = keyof typeof Jia_ll_demployeesstatuscode;

export interface Jia_ll_demployeesBase {
  importsequencenumber?: string;
  jia_costcenter?: string;
  jia_email?: string;
  jia_empid?: string;
  jia_employeestatus?: string;
  jia_ll_demployeeid: string;
  jia_name?: string;
  jia_organizationalunit?: string;
  jia_preferredname?: string;
  jia_worktype?: string;
  overriddencreatedon?: string;
  ownerid: string;
  owneridtype: string;
  statecode: Jia_ll_demployeesstatecode;
  statuscode?: Jia_ll_demployeesstatuscode;
  timezoneruleversionnumber?: string;
  utcconversiontimezonecode?: string;
}

export interface Jia_ll_demployees extends Jia_ll_demployeesBase {
  createdbyname?: string;
  createdbyyominame: string;
  createdon?: string;
  createdonbehalfbyname?: string;
  createdonbehalfbyyominame: string;
  modifiedbyname?: string;
  modifiedbyyominame: string;
  modifiedon?: string;
  modifiedonbehalfbyname?: string;
  modifiedonbehalfbyyominame: string;
  owneridname: string;
  owneridyominame: string;
  owningbusinessunitname: string;
  statecodename?: string;
  statuscodename?: string;
  versionnumber?: string;
  createdby?: object;
  _createdby_value?: string;
  createdonbehalfby?: object;
  _createdonbehalfby_value?: string;
  modifiedby?: object;
  _modifiedby_value?: string;
  modifiedonbehalfby?: object;
  _modifiedonbehalfby_value?: string;
  owningbusinessunit?: object;
  _owningbusinessunit_value?: string;
  owningteam?: object;
  _owningteam_value?: string;
  owninguser?: object;
  _owninguser_value?: string;
}