// src/generated/models/Jia_ll_dshiftplansModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */
export const Jia_ll_dshiftplansstatecode = {
  0: 'Active',
  1: 'Inactive'
} as const;
export type Jia_ll_dshiftplansstatecode = keyof typeof Jia_ll_dshiftplansstatecode;
export const Jia_ll_dshiftplansstatuscode = {
  1: 'Active',
  2: 'Inactive'
} as const;
export type Jia_ll_dshiftplansstatuscode = keyof typeof Jia_ll_dshiftplansstatuscode;

export interface Jia_ll_dshiftplansBase {
  importsequencenumber?: string;
  jia_area?: string;
  jia_colorshift?: string;
  jia_date?: string;
  jia_daynightshift?: string;
  jia_department?: string;
  jia_ll_dshiftplanid: string;
  overriddencreatedon?: string;
  ownerid: string;
  owneridtype: string;
  statecode: Jia_ll_dshiftplansstatecode;
  statuscode?: Jia_ll_dshiftplansstatuscode;
  timezoneruleversionnumber?: string;
  utcconversiontimezonecode?: string;
}

export interface Jia_ll_dshiftplans extends Jia_ll_dshiftplansBase {
  createdbyname?: string;
  createdbyyominame: string;
  createdon?: string;
  createdonbehalfbyname?: string;
  createdonbehalfbyyominame: string;
  modifiedbyname?: string;
  modifiedbyyominame: string;
  modifiedon?: string;
  modifiedonbehalfbyname?: string;
  modifiedonbehalfbyyominame: string;
  owneridname: string;
  owneridyominame: string;
  owningbusinessunitname: string;
  statecodename?: string;
  statuscodename?: string;
  versionnumber?: string;
  createdby?: object;
  _createdby_value?: string;
  createdonbehalfby?: object;
  _createdonbehalfby_value?: string;
  modifiedby?: object;
  _modifiedby_value?: string;
  modifiedonbehalfby?: object;
  _modifiedonbehalfby_value?: string;
  owningbusinessunit?: object;
  _owningbusinessunit_value?: string;
  owningteam?: object;
  _owningteam_value?: string;
  owninguser?: object;
  _owninguser_value?: string;
}