// src/generated/models/Jia_ll_fattendancereocrdsModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */
export const Jia_ll_fattendancereocrdsstatecode = {
  0: 'Active',
  1: 'Inactive'
} as const;
export type Jia_ll_fattendancereocrdsstatecode = keyof typeof Jia_ll_fattendancereocrdsstatecode;
export const Jia_ll_fattendancereocrdsstatuscode = {
  1: 'Active',
  2: 'Inactive'
} as const;
export type Jia_ll_fattendancereocrdsstatuscode = keyof typeof Jia_ll_fattendancereocrdsstatuscode;

export interface Jia_ll_fattendancereocrdsBase {
  importsequencenumber?: string;
  jia_action?: string;
  jia_area?: string;
  jia_colorshift?: string;
  jia_daynightshift?: string;
  jia_department?: string;
  jia_hours?: string;
  jia_ll_fattendancereocrdid: string;
  overriddencreatedon?: string;
  ownerid: string;
  owneridtype: string;
  statecode: Jia_ll_fattendancereocrdsstatecode;
  statuscode?: Jia_ll_fattendancereocrdsstatuscode;
  timezoneruleversionnumber?: string;
  utcconversiontimezonecode?: string;
}

export interface Jia_ll_fattendancereocrds extends Jia_ll_fattendancereocrdsBase {
  createdbyname?: string;
  createdbyyominame: string;
  createdon?: string;
  createdonbehalfbyname?: string;
  createdonbehalfbyyominame: string;
  modifiedbyname?: string;
  modifiedbyyominame: string;
  modifiedon?: string;
  modifiedonbehalfbyname?: string;
  modifiedonbehalfbyyominame: string;
  owneridname: string;
  owneridyominame: string;
  owningbusinessunitname: string;
  statecodename?: string;
  statuscodename?: string;
  versionnumber?: string;
  createdby?: object;
  _createdby_value?: string;
  createdonbehalfby?: object;
  _createdonbehalfby_value?: string;
  modifiedby?: object;
  _modifiedby_value?: string;
  modifiedonbehalfby?: object;
  _modifiedonbehalfby_value?: string;
  owningbusinessunit?: object;
  _owningbusinessunit_value?: string;
  owningteam?: object;
  _owningteam_value?: string;
  owninguser?: object;
  _owninguser_value?: string;
}